Shift serialization tests to JSONSerializationTests.swift
Previously lived in JSONSubscriptTests.swift.
